sharding, guildUpdate event for server name and avatar changes
